+++ Attributes/NodeAttribute.cs
@@ -12,6 +12,12 @@ namespace NewGraph {
         /// Color of the node as a hex string, like #FFFFFFFF. Be aware: The last two characters are for alpha values!
         /// </summary>
         public Color color = default;
+
+        /// <summary>
+        /// A custom name for the node
+        /// </summary>
+        public string nodeName = null;
+
         /// <summary>
         /// The maximum amount of allowed connections to the input port of this node.
         /// </summary>
@@ -35,14 +41,24 @@ namespace NewGraph {
         /// <param name="categories">The subcategories this node should appear under in the creation menu.</param>
         /// <param name="inputPortName">A custom name for the input port.</param>
         /// <param name="inputPortCapacity">The maximum amount of allowed connections to the input port of this node.</param>
-        public NodeAttribute(string color = null, string categories = "", string inputPortName = null, Capacity inputPortCapacity = Capacity.Multiple) {
+        public NodeAttribute(string color = null, string categories = "", string inputPortName = null, Capacity inputPortCapacity = Capacity.Multiple, string nodeName = null) {
             if (color != null) {
                 ColorUtility.TryParseHtmlString(color, out this.color);
+            }
+            if (nodeName != null) {
+                this.nodeName = nodeName;
             }
             this.inputPortName = inputPortName;
             this.inputPortCapacity = inputPortCapacity;
             this.categories = categories;
         }
 
+
+        public string GetName(Type nodeType) {
+            if (!string.IsNullOrEmpty(nodeName)) {
+                return nodeName;
+            }
+            return nodeType.Name;
+        }
     }
 }